requirements.txt
langchain[groq]
python-dotenv
langgraph
langchain
langchain[agents]
langchain-mcp-adapters
langchain[openai]
streamlit
google.genai
langchain_google_genai
langchain_google_genai
chromadb
openai
langchain_community
bs4
langchain_chroma
langgraph-checkpoint-sqlite
